--- awesome-ui/src/components/EventsTable.scss
@@ -1,3 +1,0 @@
-#scramble-btn-wrapper {
-  padding: 10px;
-}
