+++ awesome-ui/src/components/offline-scrambler/EntryInterface.jsx
