+++ awesome-ui/src/components/offline-scrambler/EntryInterface.scss
