+++ awesome-ui/src/components/offline-scrambler/EventPicker.scss
