+++ awesome-ui/src/components/offline-scrambler/OfflineScrambler.jsx
